Create mortage
A simple short with using 'if and else' program that helps the bank decide on whether or not to grant a customer the requested mortgage to buy a house.
A mortgage is accepted if the customer have been working for the same employer for more than 6 month and his salary worth at least 1% of the requested mortgage.
Two conditions should be satisfied simultaneously to have the if block executed. In such cases the "and" operator applies. The code should be as follows